requirements.txt
streamlit
pandas
numpy
google-generativeai
matplotlib